+++ glibc-malloc/linked-list-code/4-fake-node-impl.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include <stdint.h>
 
 struct node {
   int data;
@@ -164,7 +165,7 @@ int initListHeaders(Node** listHdrs, int listCount){
 }
 
 int main(void){
-  long long int listCount = 10;
+  unsigned long listCount = 10;
   Node* listHeaders[listCount*2];
   initListHeaders(listHeaders, listCount);
